Rename project from "gaid" to "pushmate"
