+++ README.md
@@ -1,3 +1,3 @@
-# gaid
+# pushmate
 
 A command line tool to automate git commits and PRs with AI.
